% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollencount.R
\name{ID}
\alias{ID}
\title{Identify a grain}
\usage{
ID(dat = info, mark = TRUE, cex = 2)
}
\arguments{
\item{dat}{the underlying locations and names of the simulated grains}

\item{mark}{Draw a green tick mark or red cross for correctly or wrongly identified grains. Set to FALSE for no drawing}

\item{cex}{size of the tick mark or cross}
}
\value{
Whether or not the grain was identified correctly
}
\description{
Identify the selected grain
}
\details{
Gives you the opportunity to interactively try and identify a specific grain.
}
