% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollencount.R
\name{select}
\alias{select}
\title{Select a grain}
\usage{
select(dat = info)
}
\arguments{
\item{dat}{the underlying locations and names of the simulated grains}
}
\value{
the type selected (invisible) and its coordinates
}
\description{
Select one of the grains simulated by the 'slide' function
}
\details{
Select one of the grains. This activates a selection feature where you click once with the pointer on one of the grains. This grain is then selected.
}
