% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollencount.R
\name{slide}
\alias{slide}
\title{Simulate a slide of pollen grains}
\usage{
slide(n = 50, dirloc = c(), size = 0.05, thumb = "100x100")
}
\arguments{
\item{n}{the number of pollen grains to simulate}

\item{dirloc}{the location of the folder with images (...). Currently the folder is filled with only a few pollen types and each folder has only a few roughly selected images of the pollen type as downloaded from various google searches. For future development, this would require more official sources.}

\item{size}{the size of the photos as drawn on the device. Keep small for slightly more realistic simulations of the real thing.}

\item{thumb}{size of the thumbnail photos}
}
\description{
Simulate a slide of pollen or other palaeoecological 'proxies', to be counted by the user
}
\details{
This is the first function to type. Simulates a slide with randomly chosen pollen grains. These are chosen from folders of photos. The same folder should have a file proportions.csv, which has two columns: the names of the types (e.g., Quercus, Pinus), and their proportions among the population of sampled pollen.
}
\examples{
slide(20)
}
